feat: account leaf index u32->u64

+++ src/circuits/balance/common/tx_settlement.rs
@@ -286,7 +286,7 @@ mod tests {
 
         let mut account_tree = AccountTree::new(ACCOUNT_TREE_HEIGHT);
         let account_leaf = AccountLeaf {
-            index: send_tree.len() as u64,
+            index: send_tree.len() as u32,
             prev: BlockNumber::default(),
             send_tree_root: send_tree.get_root(),
         };
